feat: profile of user, ui page
