+++ src/app/(profile)/user-profile/page.tsx
@@ -1,5 +1,5 @@
 import Profile from '@/components/modules/Profile/Main';
 
-export default function page() {
+export default function Page() {
   return <Profile />;
 }
