+++ src/components/modules/Profile/Right/index.tsx
@@ -1,0 +1,155 @@
+import React from 'react';
+import * as S from './style';
+export default function Right() {
+  return (
+    <>
+      <S.Right>
+        <S.MatchHistory>
+          <S.Activites>
+            <S.Content>
+              <S.UserAction>Bảo Thắng đã tham gia</S.UserAction>
+              <S.Title> Kèo anh em văn phòng đấm nhau</S.Title>
+              <S.MatchProperties>
+                <S.PropertiesList>
+                  <S.Item>
+                    <S.ItemName>Thời gian: </S.ItemName>
+                    <span>20/04/2024, 15:00</span>
+                  </S.Item>
+                  <S.Item>
+                    <S.ItemName>Địa Điểm: </S.ItemName>
+                    <span>Sân ABC, Đường XYZ</span>
+                  </S.Item>
+                  <S.Item>
+                    <S.ItemName>Cạnh tranh: </S.ItemName>
+                    <span>Vui vẻ</span>
+                  </S.Item>
+
+                  <S.Item>
+                    <S.ItemName>Khu vực: </S.ItemName>
+                    <span>Hoà Hải, Ngũ Hành Sơn, Đà Nẵng</span>
+                  </S.Item>
+                  <S.Item>
+                    <S.ItemName>Số lượng: </S.ItemName>
+                    <span>12 người</span>
+                  </S.Item>
+                  <S.Item>
+                    <S.ItemName>Tiền sân: </S.ItemName>
+                    <span>500.000 VNĐ</span>
+                  </S.Item>
+                </S.PropertiesList>
+              </S.MatchProperties>
+            </S.Content>
+          </S.Activites>
+
+          <S.Activites>
+            <S.Content>
+              <S.UserAction>Bảo Thắng đã tham gia</S.UserAction>
+              <S.Title> Kèo anh em văn phòng đấm nhau</S.Title>
+              <S.MatchProperties>
+                <S.PropertiesList>
+                  <S.Item>
+                    <S.ItemName>Thời gian: </S.ItemName>
+                    <span>20/04/2024, 15:00</span>
+                  </S.Item>
+                  <S.Item>
+                    <S.ItemName>Địa Điểm: </S.ItemName>
+                    <span>Sân ABC, Đường XYZ</span>
+                  </S.Item>
+                  <S.Item>
+                    <S.ItemName>Cạnh tranh: </S.ItemName>
+                    <span>Vui vẻ</span>
+                  </S.Item>
+
+                  <S.Item>
+                    <S.ItemName>Khu vực: </S.ItemName>
+                    <span>Hoà Hải, Ngũ Hành Sơn, Đà Nẵng</span>
+                  </S.Item>
+                  <S.Item>
+                    <S.ItemName>Số lượng: </S.ItemName>
+                    <span>12 người</span>
+                  </S.Item>
+                  <S.Item>
+                    <S.ItemName>Tiền sân: </S.ItemName>
+                    <span>500.000 VNĐ</span>
+                  </S.Item>
+                </S.PropertiesList>
+              </S.MatchProperties>
+            </S.Content>
+          </S.Activites>
+
+          <S.Activites>
+            <S.Content>
+              <S.UserAction>Bảo Thắng đã tham gia</S.UserAction>
+              <S.Title> Kèo anh em văn phòng đấm nhau</S.Title>
+              <S.MatchProperties>
+                <S.PropertiesList>
+                  <S.Item>
+                    <S.ItemName>Thời gian: </S.ItemName>
+                    <span>20/04/2024, 15:00</span>
+                  </S.Item>
+                  <S.Item>
+                    <S.ItemName>Địa Điểm: </S.ItemName>
+                    <span>Sân ABC, Đường XYZ</span>
+                  </S.Item>
+                  <S.Item>
+                    <S.ItemName>Cạnh tranh: </S.ItemName>
+                    <span>Vui vẻ</span>
+                  </S.Item>
+
+                  <S.Item>
+                    <S.ItemName>Khu vực: </S.ItemName>
+                    <span>Hoà Hải, Ngũ Hành Sơn, Đà Nẵng</span>
+                  </S.Item>
+                  <S.Item>
+                    <S.ItemName>Số lượng: </S.ItemName>
+                    <span>12 người</span>
+                  </S.Item>
+                  <S.Item>
+                    <S.ItemName>Tiền sân: </S.ItemName>
+                    <span>500.000 VNĐ</span>
+                  </S.Item>
+                </S.PropertiesList>
+              </S.MatchProperties>
+            </S.Content>
+          </S.Activites>
+
+          <S.Activites>
+            <S.Content>
+              <S.UserAction>Bảo Thắng đã tham gia</S.UserAction>
+              <S.Title> Kèo anh em văn phòng đấm nhau</S.Title>
+              <S.MatchProperties>
+                <S.PropertiesList>
+                  <S.Item>
+                    <S.ItemName>Thời gian: </S.ItemName>
+                    <span>20/04/2024, 15:00</span>
+                  </S.Item>
+                  <S.Item>
+                    <S.ItemName>Địa Điểm: </S.ItemName>
+                    <span>Sân ABC, Đường XYZ</span>
+                  </S.Item>
+                  <S.Item>
+                    <S.ItemName>Cạnh tranh: </S.ItemName>
+                    <span>Vui vẻ</span>
+                  </S.Item>
+
+                  <S.Item>
+                    <S.ItemName>Khu vực: </S.ItemName>
+                    <span>Hoà Hải, Ngũ Hành Sơn, Đà Nẵng</span>
+                  </S.Item>
+                  <S.Item>
+                    <S.ItemName>Số lượng: </S.ItemName>
+                    <span>12 người</span>
+                  </S.Item>
+                  <S.Item>
+                    <S.ItemName>Tiền sân: </S.ItemName>
+                    <span>500.000 VNĐ</span>
+                  </S.Item>
+                </S.PropertiesList>
+              </S.MatchProperties>
+            </S.Content>
+          </S.Activites>
+        </S.MatchHistory>
+      </S.Right>
+    </>
+  );
+}
